Allow for local agent installation
Remove all the OpenStack jumphost creation code since it's not necessary
for this type of deployment. Also drop some of the RHEL stuff since we're
going to be using CentOS for this deployment.

This branch is not intended to be merged into master or made into an
upstream pull request. I'll sort out some of the changes separately
and suggest changes later.